Undefined issue.
This code converts the approach from a mix of javascript / vue to the approach preffered with Pinia with the amiibo cards. Right now, everything works properly except for the fact that  the array itself is registered as undefined on instance, however it is able to be accessed through console within the same js file or the pinia menu within the vue inspect element page on the website. It is unable to be accessed through the vue file, however. I have temporarily fixed this issue by patching it with the old method of retrieving the array (which this new approach ends to replace), but it is still something that needs to be fixed nonetheless.