Code for cart is now seperated.
The code for displaying the items onto the page and for actually inserting items into the cart is now cleanly separated. Right now, the code is now split between a button and card component, a store view, and the actual javascript file that stores the functions. This makes the code much cleaner and separates all of the necessary components properly.