Comments, read description
This commit is simply an addition of comments in nearly all of the major Vue and Javascript files to give an overview to the code and how it works as practice. There are no functional or actual code changes (apart from a removal of a spare 'methods' in the 'AmiiboCard.vue' options api script). None of the project was changed.